1.添加transformer模型参数保存功能，文件格式为.model。 2.添加transformer模型读取参数功能。 3.修复adamW bug。 4.添加新rope kernel算子。

+++ src/main/java/com/omega/engine/nn/layer/LlamaMLPLayer.java
@@ -44,7 +44,7 @@ public class LlamaMLPLayer extends Layer{
 		this.oHeight = 1;
 		this.oWidth = embedDim;
 		this.nChannel = 4 * embedDim;
-		this.nChannel = 2 * nChannel / 3;
+		this.nChannel = (int)(2 * nChannel / 3);
 		this.nChannel = multiple_of * ((this.nChannel + multiple_of - 1) / multiple_of);
 		this.initLayers();
 	}
@@ -61,7 +61,7 @@ public class LlamaMLPLayer extends Layer{
 		this.oHeight = 1;
 		this.oWidth = embedDim;
 		this.nChannel = 4 * embedDim;
-		this.nChannel = 2 * nChannel / 3;
+		this.nChannel = (int)(2 * nChannel / 3);
 		this.nChannel = multiple_of * ((this.nChannel + multiple_of - 1) / multiple_of);
 		this.initLayers();
 	}
